requirements.txt
# Database connector for MySQL
mysql-connector-python>=9.1.0

# Filesystem monitoring library (event-based, not polling)
watchdog>=3.0.0

# Environment variable management
python-dotenv>=1.0.0
